requirements.txt
django==4.1
django-allauth==0.58.2
django-filter==23.4
django-oauth-toolkit==2.3.0
djangorestframework==3.13.0
drf-yasg2==1.19.4
flake8==6.1.0
isort==5.11.5
google-cloud-translate==3.12.1
numpy==1.26.2
opencv-python==4.8.1.78
pillow==10.1.0
